Revised the instructions to include a toolchain
I changed the way that the web application is being deployed.  I created a Deploy to Bluemix link that will automatically deploy the application to the person's bluemix account.  The next steps will be to connect the application to the existing Watson Machine Learning service and update the scoring endpoint in the code. This will cause the toolchain to build and deploy the application.